+++ example/DspGain.h
@@ -63,7 +63,7 @@ protected:
             _stream.assign(_stream.size(), 0);
         }
 
-        for (int i = 0; i < _stream.size(); i++)
+        for (size_t i = 0; i < _stream.size(); i++)
         {
             _stream[i] *= GetGain();
         }
